Fixes/Small changes
Fixes test errors in test_model_methods. make_data/delete_data.py changes. Creating a blog redirects to that blog. Superuser receives a warning before deleting their account.

+++ users/views.py
@@ -2,20 +2,23 @@ from django.contrib import messages
 from django.http import HttpResponseForbidden
 from django.shortcuts import render, redirect, reverse
 from django.contrib.auth import logout, get_user_model
+from senior_project.utils import login_required
 from home.models import Order, OrderHistory
 from users.forms import DeleteUserForm
 
 User = get_user_model()
 
 
+@login_required
 def profile(request):
     return render(request, 'users/profile.html')
 
 
 def delete_user(request):
+    if request.user.is_superuser:
+        return render(request, 'home/cant_delete_account.html')
     if request.user.is_anonymous or not request.user.is_authenticated:
         return HttpResponseForbidden()
-
     if request.method == 'POST':
         form = DeleteUserForm(request.POST)
